created embla carousel component
