+++ components/Component/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+const WrappedComponent = React.forwardRef(({ Component, ...props }, ref) => (
+  <Component {...props} />
+));
+
+export default WrappedComponent;
