+++ lib/marriedPeopleData.js
@@ -21,6 +21,24 @@ export const faqData = [
   },
 ];
 
+export const carouselData = [
+  {
+    title: 'Smith Auditorium',
+    image: '/marriage/slides/smith-auditorium.jpg',
+    subtitle: 'Seats 400 People',
+  },
+  {
+    title: 'Trinity Auditorium',
+    image: '/marriage/slides/smith-auditorium.jpg',
+    subtitle: 'Seats 325 People',
+  },
+  {
+    title: 'Trinity Chapel',
+    image: '/marriage/slides/smith-auditorium.jpg',
+    subtitle: 'Seats 90 People',
+  },
+];
+
 export const webflowAccordionAnimation = element => {
   // Get the accordion icon
   var siblingItems = element.parentNode.children;
